devtools/client/shared/shim/test/test_service_appinfo.html

<!DOCTYPE html>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1265802
-->
<head>
  <title>Test for Bug 1265802 - replace Services.appinfo</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css">

  <script type="application/javascript;version=1.8">
  "use strict";
  var exports = {}
  var module = {exports};
  </script>

  <script type="application/javascript;version=1.8"
	  src="resource://devtools/client/shared/shim/Services.js"></script>
</head>
<body>
<script type="application/javascript;version=1.8">
"use strict";

is(Services.appinfo.OS, SpecialPowers.Services.appinfo.OS,
   "check that Services.appinfo.OS shim matches platform");
</script>
</body>
